feat(world-gen): optimizations
