--- src/app/projects/pages/world-gen/lib/stages/tectonic-plates.ts
@@ -1,1194 +1,0 @@
-import { buildVoronoiGraph, type VoronoiEdge } from '@lib/voronoi';
-import { generateSphereVoronoi, type Vec3 } from '@lib/voronoi-sphere';
-import { OpenSimplexNoise } from '@lib/noise';
-import { mod, sphericalEmbed3D } from '@lib/math';
-import { NoiseVariables, TectonicVariables } from '../types';
-
-export enum PlateType {
-  Continental,
-  Oceanic,
-}
-
-export enum BoundaryType {
-  Convergent,
-  Divergent,
-  Transform,
-}
-
-export enum InteractionType {
-  Collision, // cont-cont convergent (Himalayas)
-  Subduction, // ocean-cont convergent (Andes)
-  OceanicConvergence, // ocean-ocean convergent (Mariana Trench)
-  ContinentalRift, // cont divergent (East Africa)
-  OceanicRidge, // ocean-ocean divergent (Mid-Atlantic Ridge)
-  Transform, // any transform (San Andreas)
-}
-
-export interface PlateProperties {
-  index: number;
-  type: PlateType;
-  dx: number;
-  dy: number;
-  /** Normalized lithospheric thickness [0, 1]. Continental ~0.55–0.85, Oceanic ~0.1–0.35. */
-  thickness: number;
-  /** Normalized crustal density [0, 1]. Continental ~0.2–0.4, Oceanic ~0.55–0.85 (higher = older). */
-  density: number;
-  /** Isostatic equilibrium base elevation, derived from thickness & density. */
-  baseElevation: number;
-}
-
-export interface BoundaryInfo {
-  plateA: number;
-  plateB: number;
-  type: BoundaryType;
-  interactionType: InteractionType;
-  /** Relative velocity magnitude, normalized to [0, 1] across all edges. */
-  intensity: number;
-  /** Shared boundary length in pixels. */
-  length: number;
-  typeA: PlateType;
-  typeB: PlateType;
-  /** Index of the plate that subducts (-1 if not applicable). */
-  subductingPlate: number;
-}
-
-export interface TectonicResult {
-  baseElevation: Float32Array;
-  faults: Float32Array;
-  mountainRanges: Float32Array;
-  plateMap: Int32Array;
-  plates: PlateProperties[];
-  boundaries: BoundaryInfo[];
-  /** Signed intra-continental perturbation (positive = swell, negative = basin). Fix B. */
-  continentalSubRelief: Float32Array;
-  /** Pixel distance to nearest ridge/rift seed (Chamfer). Feeds the ocean age gradient. Fix C. */
-  distToRidge: Float32Array;
-  /** distToRidge normalized to [0, 1] (1 = farthest abyssal plain from any ridge). Fix C. */
-  oceanAge: Float32Array;
-}
-
-/** Simple seeded PRNG (xorshift32). */
-function xorshift32(state: { s: number }): number {
-  let s = state.s;
-  s ^= s << 13;
-  s ^= s >>> 17;
-  s ^= s << 5;
-  state.s = s;
-  return (s >>> 0) / 0xffffffff;
-}
-
-// ── Cell grouping ────────────────────────────────────────────────────────
-
-/** Update minDist array with distances from a newly placed seed. */
-function updateMinDist(
-  seeds: { x: number; y: number }[],
-  minDist: Float32Array,
-  newSeed: number,
-  cellCount: number,
-  width: number
-): void {
-  const sx = seeds[newSeed].x;
-  const sy = seeds[newSeed].y;
-  for (let i = 0; i < cellCount; i++) {
-    let dx = seeds[i].x - sx;
-    if (dx > width / 2) dx -= width;
-    else if (dx < -width / 2) dx += width;
-    const dy = seeds[i].y - sy;
-    const d = dx * dx + dy * dy;
-    if (d < minDist[i]) minDist[i] = d;
-  }
-}
-
-/**
- * Two-phase seed placement:
- *
- * Phase 1 — Major plates (largest ~40%) are placed via farthest-point sampling,
- * ensuring they're well-spaced across the map.
- *
- * Phase 2 — A preview Dijkstra partitions the cell graph using only the major
- * plates. The actual computed boundaries from this partition are identified.
- * Minor plate seeds are then placed at these boundary cells via farthest-point
- * sampling. This guarantees minor plates always spawn at the margins between
- * major plates, never as islands inside them.
- */
-function selectPlateSeeds(
-  seeds: { x: number; y: number }[],
-  cellCount: number,
-  plateCount: number,
-  width: number,
-  targetFractions: number[],
-  neighbors: number[][],
-  rng: { s: number }
-): number[] {
-  const plateSeeds = new Array<number>(plateCount);
-  const placed = new Uint8Array(cellCount);
-
-  // Sort plates by target fraction descending → major plates first
-  const order = targetFractions.map((f, i) => ({ fraction: f, plate: i })).sort((a, b) => b.fraction - a.fraction);
-
-  const majorCount = Math.max(3, Math.ceil(plateCount * 0.4));
-
-  // --- Phase 1: Major plates via farthest-point sampling ---
-  const minDist = new Float32Array(cellCount).fill(Infinity);
-
-  const first = Math.floor(xorshift32(rng) * cellCount);
-  plateSeeds[order[0].plate] = first;
-  placed[first] = 1;
-
-  for (let m = 1; m < majorCount; m++) {
-    updateMinDist(seeds, minDist, plateSeeds[order[m - 1].plate], cellCount, width);
-
-    let best = -1;
-    let bestD = -1;
-    for (let i = 0; i < cellCount; i++) {
-      if (!placed[i] && minDist[i] > bestD) {
-        bestD = minDist[i];
-        best = i;
-      }
-    }
-    plateSeeds[order[m].plate] = best;
-    placed[best] = 1;
-  }
-  updateMinDist(seeds, minDist, plateSeeds[order[majorCount - 1].plate], cellCount, width);
-
-  // --- Phase 2: Preview Dijkstra with major plates to find real boundaries ---
-  const previewSeeds: number[] = [];
-  const previewFractions: number[] = [];
-  for (let m = 0; m < majorCount; m++) {
-    previewSeeds.push(plateSeeds[order[m].plate]);
-    previewFractions.push(order[m].fraction);
-  }
-  const previewAssignment = groupCellsIntoPlates(neighbors, cellCount, majorCount, previewSeeds, previewFractions);
-
-  // Mark cells at boundaries of the preview partition
-  const isBoundaryCell = new Uint8Array(cellCount);
-  for (let c = 0; c < cellCount; c++) {
-    for (const n of neighbors[c]) {
-      if (previewAssignment[n] !== previewAssignment[c]) {
-        isBoundaryCell[c] = 1;
-        break;
-      }
-    }
-  }
-
-  // --- Phase 3: Minor seeds at boundary cells via farthest-point sampling ---
-  for (let m = majorCount; m < plateCount; m++) {
-    let best = -1;
-    let bestD = -1;
-
-    // Pick the boundary cell farthest from all placed seeds
-    for (let i = 0; i < cellCount; i++) {
-      if (!placed[i] && isBoundaryCell[i] && minDist[i] > bestD) {
-        bestD = minDist[i];
-        best = i;
-      }
-    }
-
-    if (best === -1) {
-      // Fallback: any unplaced cell (boundary cells exhausted)
-      for (let i = 0; i < cellCount; i++) {
-        if (!placed[i] && minDist[i] > bestD) {
-          bestD = minDist[i];
-          best = i;
-        }
-      }
-    }
-
-    plateSeeds[order[m].plate] = best;
-    placed[best] = 1;
-    updateMinDist(seeds, minDist, best, cellCount, width);
-  }
-
-  return plateSeeds;
-}
-
-/**
- * Generate target area fractions using a power-law distribution.
- * A few large plates cover most of the globe, many small plates fill the rest.
- * The fractions are shuffled so large plates aren't always the first seeds.
- */
-function generateSizeDistribution(plateCount: number, rng: { s: number }): number[] {
-  const ALPHA = 1.2;
-  const raw: number[] = [];
-  for (let i = 0; i < plateCount; i++) {
-    raw.push(1 / Math.pow(i + 1, ALPHA));
-  }
-  const sum = raw.reduce((a, b) => a + b, 0);
-  const fractions = raw.map((v) => v / sum);
-
-  // Fisher-Yates shuffle so large plates aren't always the first seeds
-  for (let i = fractions.length - 1; i > 0; i--) {
-    const j = Math.floor(xorshift32(rng) * (i + 1));
-    [fractions[i], fractions[j]] = [fractions[j], fractions[i]];
-  }
-
-  return fractions;
-}
-
-/**
- * Group voronoi cells into plates using cost-scaled multi-source Dijkstra.
- *
- * All plates start growing simultaneously from their seeds, but each plate
- * has a per-hop cost derived from its target area fraction. Large plates
- * have low cost (grow fast / expand far), small plates have high cost
- * (grow slow / stay small). The competitive growth naturally partitions
- * the cell graph with sizes matching the target distribution.
- */
-function groupCellsIntoPlates(
-  neighbors: number[][],
-  cellCount: number,
-  plateCount: number,
-  plateSeeds: number[],
-  targetFractions: number[]
-): Int32Array {
-  const cellToPlate = new Int32Array(cellCount).fill(-1);
-
-  // Compute per-plate growth cost: cost ∝ 1/sqrt(fraction).
-  // Large plates have low cost (grow far), small plates have high cost (grow short).
-  // In competitive 2D growth, area ∝ radius² and radius ∝ 1/cost,
-  // so area ∝ 1/cost² ∝ fraction. Cap to avoid extreme ratios.
-  const maxFraction = Math.max(...targetFractions);
-  const costs = targetFractions.map((f) => Math.min(Math.sqrt(maxFraction / f), 10));
-
-  // Min-heap priority queue
-  const heap: { priority: number; cell: number; plate: number }[] = [];
-
-  const push = (priority: number, cell: number, plate: number) => {
-    heap.push({ priority, cell, plate });
-    let i = heap.length - 1;
-    while (i > 0) {
-      const parent = (i - 1) >> 1;
-      if (heap[i].priority >= heap[parent].priority) break;
-      [heap[i], heap[parent]] = [heap[parent], heap[i]];
-      i = parent;
-    }
-  };
-
-  const pop = () => {
-    const top = heap[0];
-    const last = heap.pop()!;
-    if (heap.length > 0) {
-      heap[0] = last;
-      let i = 0;
-      while (true) {
-        let smallest = i;
-        const l = 2 * i + 1;
-        const r = 2 * i + 2;
-        if (l < heap.length && heap[l].priority < heap[smallest].priority) smallest = l;
-        if (r < heap.length && heap[r].priority < heap[smallest].priority) smallest = r;
-        if (smallest === i) break;
-        [heap[i], heap[smallest]] = [heap[smallest], heap[i]];
-        i = smallest;
-      }
-    }
-    return top;
-  };
-
-  // Pre-claim all seed cells
-  for (let p = 0; p < plateCount; p++) {
-    cellToPlate[plateSeeds[p]] = p;
-  }
-
-  // Seed the queue with neighbors of all plate seeds (all start simultaneously)
-  for (let p = 0; p < plateCount; p++) {
-    for (const n of neighbors[plateSeeds[p]]) {
-      if (cellToPlate[n] === -1) {
-        push(costs[p], n, p);
-      }
-    }
-  }
-
-  // Dijkstra: claim cells in priority order, each hop costs costs[plate]
-  while (heap.length > 0) {
-    const { priority, cell, plate } = pop();
-    if (cellToPlate[cell] !== -1) continue; // already claimed
-
-    cellToPlate[cell] = plate;
-
-    for (const n of neighbors[cell]) {
-      if (cellToPlate[n] === -1) {
-        push(priority + costs[plate], n, plate);
-      }
-    }
-  }
-
-  // Safety: assign any remaining unclaimed cells to a neighbor's plate
-  for (let i = 0; i < cellCount; i++) {
-    if (cellToPlate[i] === -1) {
-      for (const n of neighbors[i]) {
-        if (cellToPlate[n] !== -1) {
-          cellToPlate[i] = cellToPlate[n];
-          break;
-        }
-      }
-    }
-  }
-
-  return cellToPlate;
-}
-
-// ── Plate properties ─────────────────────────────────────────────────────
-
-interface PlateCentroid {
-  x: number;
-  y: number;
-}
-
-/**
- * Project unit-sphere seed positions to equirectangular pixel coords:
- * lon = atan2(y, x), lat = asin(z); pxX = (lon+π)/(2π)·W, pxY = (π/2−lat)/π·H.
- */
-function sphereSeedsToPixels(sphereSeeds: Vec3[], width: number, height: number): { x: number; y: number }[] {
-  const TWO_PI = Math.PI * 2;
-  const out: { x: number; y: number }[] = new Array(sphereSeeds.length);
-  for (let i = 0; i < sphereSeeds.length; i++) {
-    const s = sphereSeeds[i];
-    const lat = Math.asin(s.z < -1 ? -1 : s.z > 1 ? 1 : s.z);
-    const lon = Math.atan2(s.y, s.x);
-    out[i] = {
-      x: ((lon + Math.PI) / TWO_PI) * width,
-      y: ((Math.PI / 2 - lat) / Math.PI) * height,
-    };
-  }
-  return out;
-}
-
-/**
- * Compute plate centroids on the sphere: sum the unit-vector positions of all
- * member-cell seeds, renormalize to the unit sphere, then project to pixel
- * coords. This is the natural sphere centroid (no pole-bias from cylindrical
- * y-averaging) and is well-defined even for plates that span the antimeridian.
- */
-function computePlateCentroidsSphere(
-  sphereSeeds: Vec3[],
-  cellToPlate: Int32Array,
-  cellCount: number,
-  plateCount: number,
-  width: number,
-  height: number
-): PlateCentroid[] {
-  const sumX = new Float64Array(plateCount);
-  const sumY = new Float64Array(plateCount);
-  const sumZ = new Float64Array(plateCount);
-  const counts = new Int32Array(plateCount);
-
-  for (let i = 0; i < cellCount; i++) {
-    const p = cellToPlate[i];
-    if (p < 0) continue;
-    counts[p]++;
-    const s = sphereSeeds[i];
-    sumX[p] += s.x;
-    sumY[p] += s.y;
-    sumZ[p] += s.z;
-  }
-
-  const TWO_PI = Math.PI * 2;
-  const centroids: PlateCentroid[] = [];
-  for (let p = 0; p < plateCount; p++) {
-    if (counts[p] === 0) {
-      centroids.push({ x: 0, y: 0 });
-      continue;
-    }
-    const cx = sumX[p];
-    const cy = sumY[p];
-    const cz = sumZ[p];
-    const len = Math.sqrt(cx * cx + cy * cy + cz * cz);
-    if (len === 0) {
-      centroids.push({ x: 0, y: 0 });
-      continue;
-    }
-    const nx = cx / len;
-    const ny = cy / len;
-    const nz = cz / len;
-    const lat = Math.asin(nz < -1 ? -1 : nz > 1 ? 1 : nz);
-    const lon = Math.atan2(ny, nx);
-    centroids.push({
-      x: mod(((lon + Math.PI) / TWO_PI) * width, width),
-      y: ((Math.PI / 2 - lat) / Math.PI) * height,
-    });
-  }
-
-  return centroids;
-}
-
-/**
- * Compute isostatic equilibrium base elevation from thickness and density.
- *
- * Thicker, less-dense crust floats higher on the mantle (continental highlands).
- * Thinner, denser crust sits lower (oceanic basins).
- * Calibrated so average continental ≈ 0.25, average oceanic ≈ -0.35.
- */
-export function computeIsostasy(thickness: number, density: number): number {
-  return 1.5 * thickness * (1 - density) - 0.5;
-}
-
-/**
- * Assign plate type and drift using a low-frequency noise field.
- * Sampling noise at each plate's centroid creates spatial coherence.
- * Thickness and density are derived from plate type with random variation,
- * representing geological age and composition differences.
- */
-function assignPlateProperties(
-  centroids: PlateCentroid[],
-  plateCount: number,
-  width: number,
-  height: number,
-  seed: number
-): PlateProperties[] {
-  const rng = { s: (seed ^ 0xdeadbeef) | 1 };
-  const continentNoise = new OpenSimplexNoise((seed ^ 0xcafebeef) | 0);
-  // The unit-vector embedding confines noise samples to the box [−1, 1]³, so
-  // freq must be high enough that multiple distinct features fit on the
-  // sphere. Below ~1.0 a single noise blob dominates the whole globe and its
-  // gradient along z = sin(lat) forces a hemispheric bias (one whole hemi
-  // continental, the other oceanic). 1.6 gives ~3 features pole-to-pole and
-  // ~1.6 around the equator — continent-sized chunks at all latitudes.
-  const CONTINENT_FREQ = 1.6;
-
-  const np = new Float32Array(3);
-  const plates: PlateProperties[] = [];
-  for (let i = 0; i < plateCount; i++) {
-    sphericalEmbed3D(centroids[i].x, centroids[i].y, width, height, np);
-    const sample = continentNoise.eval3D(np[0] * CONTINENT_FREQ, np[1] * CONTINENT_FREQ, np[2] * CONTINENT_FREQ);
-
-    const angle = xorshift32(rng) * Math.PI * 2;
-    const magnitude = 0.3 + xorshift32(rng) * 0.7;
-
-    const type = sample > 0 ? PlateType.Continental : PlateType.Oceanic;
-
-    // Thickness & density vary by plate type with random spread
-    let thickness: number;
-    let density: number;
-    if (type === PlateType.Continental) {
-      // Continental: thick (0.55–0.85), low density (0.2–0.4)
-      thickness = 0.55 + xorshift32(rng) * 0.3;
-      density = 0.2 + xorshift32(rng) * 0.2;
-    } else {
-      // Oceanic: thin (0.1–0.35), higher density (0.55–0.85)
-      // Higher density represents older, cooler oceanic lithosphere
-      thickness = 0.1 + xorshift32(rng) * 0.25;
-      density = 0.55 + xorshift32(rng) * 0.3;
-    }
-
-    plates.push({
-      index: i,
-      type,
-      dx: Math.cos(angle) * magnitude,
-      dy: Math.sin(angle) * magnitude,
-      thickness,
-      density,
-      baseElevation: computeIsostasy(thickness, density),
-    });
-  }
-
-  return plates;
-}
-
-// ── Boundary aggregation & classification ─────────────────────────────────
-
-/**
- * Aggregate cell-level voronoi edges into plate-level boundaries.
- * Only edges between cells belonging to different plates are included.
- */
-function aggregatePlateBoundaries(
-  cellEdges: VoronoiEdge[],
-  cellToPlate: Int32Array,
-  plateCount: number
-): { plateA: number; plateB: number; length: number }[] {
-  const edgeMap = new Map<number, number>(); // packed key -> index
-  const plateBoundaries: { plateA: number; plateB: number; length: number }[] = [];
-
-  for (const edge of cellEdges) {
-    const pA = cellToPlate[edge.cellA];
-    const pB = cellToPlate[edge.cellB];
-    if (pA === pB) continue; // intra-plate edge
-
-    const lo = Math.min(pA, pB);
-    const hi = Math.max(pA, pB);
-    const key = lo * plateCount + hi;
-
-    const idx = edgeMap.get(key);
-    if (idx !== undefined) {
-      plateBoundaries[idx].length += edge.length;
-    } else {
-      edgeMap.set(key, plateBoundaries.length);
-      plateBoundaries.push({ plateA: lo, plateB: hi, length: edge.length });
-    }
-  }
-
-  return plateBoundaries;
-}
-
-function deriveInteractionType(type: BoundaryType, typeA: PlateType, typeB: PlateType): InteractionType {
-  const bothContinental = typeA === PlateType.Continental && typeB === PlateType.Continental;
-  const bothOceanic = typeA === PlateType.Oceanic && typeB === PlateType.Oceanic;
-
-  switch (type) {
-    case BoundaryType.Convergent:
-      if (bothContinental) return InteractionType.Collision;
-      if (bothOceanic) return InteractionType.OceanicConvergence;
-      return InteractionType.Subduction;
-    case BoundaryType.Divergent:
-      if (bothOceanic) return InteractionType.OceanicRidge;
-      return InteractionType.ContinentalRift;
-    case BoundaryType.Transform:
-      return InteractionType.Transform;
-  }
-}
-
-function classifyBoundaries(
-  rawBoundaries: { plateA: number; plateB: number; length: number }[],
-  plates: PlateProperties[],
-  centroids: PlateCentroid[],
-  width: number,
-  height: number
-): BoundaryInfo[] {
-  const classified: { boundary: (typeof rawBoundaries)[0]; type: BoundaryType; rawIntensity: number }[] = [];
-  let maxIntensity = 0;
-
-  for (const boundary of rawBoundaries) {
-    const pA = plates[boundary.plateA];
-    const pB = plates[boundary.plateB];
-    const cA = centroids[boundary.plateA];
-    const cB = centroids[boundary.plateB];
-
-    // Wrap-aware A→B vector. Scale x components by cos(lat) at the boundary
-    // midpoint so geometry is in km-equivalent units rather than pixels — at
-    // high latitudes one pixel of x covers far less ground than one pixel of y.
-    const midY = (cA.y + cB.y) * 0.5;
-    const midLat = Math.PI / 2 - ((midY + 0.5) / height) * Math.PI;
-    const cosMid = Math.cos(midLat);
-
-    let nxPx = cB.x - cA.x;
-    if (nxPx > width / 2) nxPx -= width;
-    else if (nxPx < -width / 2) nxPx += width;
-    const nx = nxPx * cosMid;
-    const ny = cB.y - cA.y;
-    const nLen = Math.sqrt(nx * nx + ny * ny);
-    if (nLen === 0) continue;
-    const ux = nx / nLen;
-    const uy = ny / nLen;
-
-    // Relative drift of A toward B (same cos(lat) scaling on x).
-    const relDx = (pA.dx - pB.dx) * cosMid;
-    const relDy = pA.dy - pB.dy;
-
-    // Project onto normal (positive = convergent)
-    const vNormal = relDx * ux + relDy * uy;
-    const vTanX = relDx - vNormal * ux;
-    const vTanY = relDy - vNormal * uy;
-    const vTangent = Math.sqrt(vTanX * vTanX + vTanY * vTanY);
-
-    let type: BoundaryType;
-    if (Math.abs(vNormal) > vTangent) {
-      type = vNormal > 0 ? BoundaryType.Convergent : BoundaryType.Divergent;
-    } else {
-      type = BoundaryType.Transform;
-    }
-
-    const rawIntensity = Math.sqrt(relDx * relDx + relDy * relDy);
-    if (rawIntensity > maxIntensity) maxIntensity = rawIntensity;
-
-    classified.push({ boundary, type, rawIntensity });
-  }
-
-  const invMax = maxIntensity > 0 ? 1 / maxIntensity : 1;
-  return classified.map(({ boundary, type, rawIntensity }) => {
-    const pA = plates[boundary.plateA];
-    const pB = plates[boundary.plateB];
-    const interactionType = deriveInteractionType(type, pA.type, pB.type);
-
-    // Determine which plate subducts based on density
-    let subductingPlate = -1;
-    if (interactionType === InteractionType.Subduction) {
-      // In ocean-continent convergence, the oceanic plate always subducts
-      subductingPlate = pA.type === PlateType.Oceanic ? boundary.plateA : boundary.plateB;
-    } else if (interactionType === InteractionType.OceanicConvergence) {
-      // In ocean-ocean convergence, the denser (older) plate subducts
-      subductingPlate = pA.density >= pB.density ? boundary.plateA : boundary.plateB;
-    }
-
-    return {
-      plateA: boundary.plateA,
-      plateB: boundary.plateB,
-      type,
-      interactionType,
-      intensity: rawIntensity * invMax,
-      length: boundary.length,
-      typeA: pA.type,
-      typeB: pB.type,
-      subductingPlate,
-    };
-  });
-}
-
-// ── Interaction-based elevation profiles ──────────────────────────────────
-
-/**
- * Compute the elevation delta and mountain range intensity for a pixel
- * near a plate boundary, based on the physical interaction type.
- *
- * `t` is the normalized falloff (1.0 at the boundary, 0.0 at falloff edge).
- * `pixelPlateType` is the plate type of the pixel being evaluated —
- * this matters for asymmetric interactions like subduction.
- * `isSubductingSide` indicates whether this pixel belongs to the subducting plate —
- * used for asymmetric elevation in convergent boundaries.
- */
-export function interactionElevation(
-  interaction: InteractionType,
-  pixelPlateType: PlateType,
-  intensity: number,
-  t: number,
-  isSubductingSide = false
-): { elevDelta: number; mountainRange: number } {
-  const f = intensity * t;
-  switch (interaction) {
-    case InteractionType.Collision:
-      return { elevDelta: 0.4 * f, mountainRange: 0.8 * t };
-    case InteractionType.Subduction:
-      if (pixelPlateType === PlateType.Continental) {
-        return { elevDelta: 0.35 * f, mountainRange: 0.6 * t };
-      }
-      return { elevDelta: -0.15 * f, mountainRange: 0 };
-    case InteractionType.OceanicConvergence:
-      // Denser plate subducts → trench; lighter plate gets volcanic island arc
-      if (isSubductingSide) {
-        return { elevDelta: -0.12 * f, mountainRange: 0 };
-      }
-      return { elevDelta: 0.1 * f, mountainRange: 0.3 * t };
-    case InteractionType.ContinentalRift:
-      return { elevDelta: -0.1 * f, mountainRange: 0 };
-    case InteractionType.OceanicRidge:
-      return { elevDelta: 0.03 * f, mountainRange: 0 };
-    case InteractionType.Transform:
-      return { elevDelta: 0, mountainRange: 0 };
-  }
-}
-
-// ── Per-interaction falloff width multipliers ────────────────────────────
-
-/**
- * Falloff width multiplier per interaction type.
- * Collision zones (Himalayas) are wide; transform faults (San Andreas) are narrow.
- * For subduction, the overriding plate side is wider than the trench side.
- */
-const FALLOFF_MULTIPLIER: Record<InteractionType, number> = {
-  [InteractionType.Collision]: 1.8,
-  [InteractionType.Subduction]: 1.4, // overriding (continental) side
-  [InteractionType.OceanicConvergence]: 0.7,
-  [InteractionType.ContinentalRift]: 1.2,
-  [InteractionType.OceanicRidge]: 0.8,
-  [InteractionType.Transform]: 0.35,
-};
-
-/** Subducting-side falloff is narrower (trench). */
-const SUBDUCTION_TRENCH_MULTIPLIER = 0.4;
-
-// ── Continental shelf ─────────────────────────────────────────────────────
-
-/**
- * Width of the continental shelf band (in multiples of baseFalloff), measured
- * inland from a passive-style continental-oceanic margin. Scales with map size
- * via baseFalloff.
- */
-const SHELF_WIDTH_MULTIPLIER = 3.0;
-
-/**
- * Target elevation for the near-margin shelf band. Sits just below the sea
- * level the percentile rule tends to land on, so fBm detail can lift parts
- * of it back above water as offshore islands.
- */
-const SHELF_TARGET = -0.25;
-
-// ── Rasterization ─────────────────────────────────────────────────────────
-
-function rasterizePlateInteractions(
-  plateMap: Int32Array,
-  boundaries: BoundaryInfo[],
-  plates: PlateProperties[],
-  width: number,
-  height: number,
-  nv: NoiseVariables,
-  tv: TectonicVariables
-): {
-  baseElevation: Float32Array;
-  faults: Float32Array;
-  mountainRanges: Float32Array;
-  continentalSubRelief: Float32Array;
-  distToRidge: Float32Array;
-  oceanAge: Float32Array;
-} {
-  const falloffScale = tv.boundaryFalloffScale;
-  const size = width * height;
-  const baseElevation = new Float32Array(size);
-  const faults = new Float32Array(size);
-  const mountainRanges = new Float32Array(size);
-
-  const plateCount = plates.length;
-
-  // Build boundary lookup: packed key -> boundary array index
-  const boundaryMap = new Map<number, number>();
-  for (let bi = 0; bi < boundaries.length; bi++) {
-    const b = boundaries[bi];
-    const lo = Math.min(b.plateA, b.plateB);
-    const hi = Math.max(b.plateA, b.plateB);
-    boundaryMap.set(lo * plateCount + hi, bi);
-  }
-
-  // Pass 1: Detect plate-boundary pixels, record highest-intensity boundary index
-  const isBoundary = new Uint8Array(size);
-  const nearestBoundary = new Int16Array(size).fill(-1);
-  // Pixels on continental plates that border an oceanic plate via a passive-style
-  // interaction (transform or rift). Seeds the shelf distance transform so the
-  // Jacobi target can sag toward shallow-water depth on this side of the margin.
-  // Subduction continental-side pixels are intentionally excluded so the
-  // mountain belt there keeps its full height.
-  const shelfSeedMask = new Uint8Array(size);
-
-  for (let y = 0; y < height; y++) {
-    for (let x = 0; x < width; x++) {
-      const idx = y * width + x;
-      const plate = plateMap[idx];
-
-      const neighbors = [
-        x > 0 ? plateMap[idx - 1] : plateMap[y * width + width - 1],
-        x < width - 1 ? plateMap[idx + 1] : plateMap[y * width],
-        y > 0 ? plateMap[idx - width] : -1,
-        y < height - 1 ? plateMap[idx + width] : -1,
-      ];
-
-      let bestIntensity = -1;
-      let hasPassiveOceanicNeighbor = false;
-      for (const neighbor of neighbors) {
-        if (neighbor >= 0 && neighbor !== plate) {
-          const lo = Math.min(plate, neighbor);
-          const hi = Math.max(plate, neighbor);
-          const bi = boundaryMap.get(lo * plateCount + hi);
-          if (bi !== undefined) {
-            const nb = boundaries[bi];
-            if (nb.intensity > bestIntensity) {
-              bestIntensity = nb.intensity;
-              nearestBoundary[idx] = bi;
-            }
-            if (
-              plates[plate].type === PlateType.Continental &&
-              plates[neighbor].type === PlateType.Oceanic &&
-              (nb.interactionType === InteractionType.Transform ||
-                nb.interactionType === InteractionType.ContinentalRift)
-            ) {
-              hasPassiveOceanicNeighbor = true;
-            }
-          }
-        }
-      }
-      if (bestIntensity >= 0) {
-        isBoundary[idx] = 1;
-      }
-      if (hasPassiveOceanicNeighbor) {
-        shelfSeedMask[idx] = 1;
-      }
-    }
-  }
-
-  // Pass 2: Chamfer distance transform
-  const dist = chamferDistance(isBoundary, width, height);
-
-  // Pass 2b: Chamfer distance to nearest passive-margin shelf seed. Continental
-  // pixels within shelf range get their gravity target pulled toward SHELF_TARGET,
-  // producing the shallow band that gives coastlines their ragged, island-dotted
-  // character. Pixels far inland (or outside continental plates entirely) keep
-  // their plate's baseline as the target.
-  const distToShelfSeed = chamferDistance(shelfSeedMask, width, height);
-
-  // Pass 3: BFS from boundary pixels to propagate boundary index outward
-  // Use the maximum possible falloff width for BFS reach, then per-pixel widths in Pass 5
-  const baseFalloff = Math.max(4, Math.round(width / 64));
-  const maxFalloffWidth = Math.ceil(baseFalloff * 1.8 * falloffScale); // 1.8 = largest multiplier (Collision)
-  propagateBoundaryIndex(nearestBoundary, isBoundary, dist, maxFalloffWidth, width, height);
-
-  const shelfWidth = baseFalloff * SHELF_WIDTH_MULTIPLIER * falloffScale;
-
-  // Pass 4: Jacobi relaxation for smooth interior elevation gradients.
-  // Each pixel has an "effective base" — usually its plate's isostatic
-  // baseline, but pulled toward SHELF_TARGET on the continental side of
-  // passive-style cont-ocean margins.
-  const effectiveBase = (idx: number): number => {
-    const plate = plates[plateMap[idx]];
-    if (plate.type !== PlateType.Continental) return plate.baseElevation;
-    const d = distToShelfSeed[idx];
-    if (d >= shelfWidth) return plate.baseElevation;
-    const t = d / shelfWidth; // 0 at seed, 1 at far edge of shelf band
-    return SHELF_TARGET * (1 - t) + plate.baseElevation * t;
-  };
-
-  // Initialize every pixel (boundary and interior) at its effective base.
-  // This lets continental boundary pixels adjacent to passive oceanic margins
-  // sag toward SHELF_TARGET along with their neighbors, instead of standing up
-  // as a one-pixel cliff above the shelf (which previously formed a ring of
-  // strip-islands tracing the continental fault lines).
-  for (let i = 0; i < size; i++) {
-    baseElevation[i] = effectiveBase(i);
-  }
-
-  // Jacobi iterations: smooth interior while keeping boundaries pinned
-  const JACOBI_ITERATIONS = 8;
-  const GRAVITY_ALPHA = 0.1; // pull toward effective base — higher = flatter interiors
-  const temp = new Float32Array(size);
-
-  for (let iter = 0; iter < JACOBI_ITERATIONS; iter++) {
-    for (let y = 0; y < height; y++) {
-      for (let x = 0; x < width; x++) {
-        const idx = y * width + x;
-
-        // Average of 4 neighbors (wrap x, clamp y) — needed for both pinned
-        // (soft-blend) and interior branches under Fix D.
-        const left = baseElevation[y * width + mod(x - 1, width)];
-        const right = baseElevation[y * width + mod(x + 1, width)];
-        const up = y > 0 ? baseElevation[(y - 1) * width + x] : baseElevation[idx];
-        const down = y < height - 1 ? baseElevation[(y + 1) * width + x] : baseElevation[idx];
-        const avg = (left + right + up + down) / 4;
-
-        if (isBoundary[idx]) {
-          // Fix D: passive seams (transforms, continental rifts) blend halfway
-          // with neighbors so they don't survive smoothing as one-pixel cliffs.
-          // Orogenic boundaries stay hard-pinned so mountain belts keep their crest.
-          if (tv.boundarySoftenPassive) {
-            const bi = nearestBoundary[idx];
-            const type = bi >= 0 ? boundaries[bi].interactionType : null;
-            if (type === InteractionType.Transform || type === InteractionType.ContinentalRift) {
-              temp[idx] = 0.5 * baseElevation[idx] + 0.5 * avg;
-              continue;
-            }
-          }
-          temp[idx] = baseElevation[idx]; // pinned
-          continue;
-        }
-
-        temp[idx] = (1 - GRAVITY_ALPHA) * avg + GRAVITY_ALPHA * effectiveBase(idx);
-      }
-    }
-    // Swap
-    baseElevation.set(temp);
-  }
-
-  // Pass 4.5: Intra-plate relief so coastlines aren't uniform-by-plate.
-  //   - Continental pixels get a signed ridged-fBm perturbation — cratons,
-  //     swells, intracratonic basins (Hudson-Bay-style embayments).
-  //   - Oceanic pixels get an age-from-ridge exponential gradient —
-  //     mid-ocean ridges rise, abyssal plains sink.
-  // Applied *after* Jacobi (so smoothing doesn't erase it) but *before* the
-  // interaction-delta overlay (so orogenic belts still sit on top).
-  const continentalSubRelief = new Float32Array(size);
-
-  // Seed the ridge-distance field from divergent-boundary pixels only.
-  const ridgeSeedMask = new Uint8Array(size);
-  let hasRidgeSeeds = false;
-  for (let i = 0; i < size; i++) {
-    if (!isBoundary[i]) continue;
-    const bi = nearestBoundary[i];
-    if (bi < 0) continue;
-    const type = boundaries[bi].interactionType;
-    if (type === InteractionType.OceanicRidge || type === InteractionType.ContinentalRift) {
-      ridgeSeedMask[i] = 1;
-      hasRidgeSeeds = true;
-    }
-  }
-  const distToRidge = chamferDistance(ridgeSeedMask, width, height);
-  const oceanAge = new Float32Array(size);
-  const oceanAgeScale = width / 6;
-
-  // Find the max *finite* ridge distance for oceanAge layer normalization.
-  // chamferDistance fills unreachable cells with INF = width+height; when
-  // no ridge seeds exist, every cell stays at INF and the normalized field
-  // is left at 0.
-  if (hasRidgeSeeds) {
-    const INF_SENTINEL = width + height - 1;
-    let maxRidgeDist = 0;
-    for (let i = 0; i < size; i++) {
-      const d = distToRidge[i];
-      if (d > maxRidgeDist && d <= INF_SENTINEL) maxRidgeDist = d;
-    }
-    const invMax = maxRidgeDist > 0 ? 1 / maxRidgeDist : 0;
-    for (let i = 0; i < size; i++) {
-      oceanAge[i] = Math.min(1, distToRidge[i] * invMax);
-    }
-  }
-
-  const subReliefNoise = new OpenSimplexNoise((nv.seed ^ 0xbee5ca7e) | 0);
-  const subReliefFreq = nv.frequency * tv.continentalSubReliefFrequency;
-  const subReliefAmp = tv.continentalSubReliefAmplitude;
-  const ageStrength = tv.oceanicAgeGradientStrength;
-
-  const npSub = new Float32Array(3);
-  for (let y = 0; y < height; y++) {
-    for (let x = 0; x < width; x++) {
-      const idx = y * width + x;
-      const plate = plates[plateMap[idx]];
-      if (plate.type === PlateType.Continental) {
-        sphericalEmbed3D(x, y, width, height, npSub);
-        // Plain fBm in [-1, 1], symmetric around 0 → roughly half swells
-        // (positive) and half basins (negative). The earlier ridged-fBm form
-        // (1 − |n|) only produced positive deltas, so the diverging red/blue
-        // sub-relief layer never showed blue.
-        let sum = 0;
-        let range = 0;
-        let f = subReliefFreq;
-        let a = 1;
-        for (let o = 0; o < 4; o++) {
-          const n = subReliefNoise.eval3D(npSub[0] * f, npSub[1] * f, npSub[2] * f);
-          sum += n * a;
-          range += a;
-          f *= 2.0;
-          a *= 0.5;
-        }
-        const signed = sum / Math.max(range, 1);
-        const delta = signed * subReliefAmp;
-        continentalSubRelief[idx] = delta;
-        baseElevation[idx] += delta;
-      } else if (hasRidgeSeeds) {
-        // Oceanic age-from-ridge gradient: ridges lift, abyssal plains sink.
-        const d = distToRidge[idx];
-        const rawDelta = 0.15 * Math.exp(-d / oceanAgeScale) - 0.1;
-        baseElevation[idx] += rawDelta * ageStrength;
-      }
-    }
-  }
-
-  // Pass 5: Add interaction falloff for faults and mountainRanges,
-  // and overlay interaction elevation delta on top of the gradient base.
-  // Each interaction type has its own falloff width; subduction is asymmetric.
-  //
-  // `dist[i]` is warped by a low-amplitude noise before the falloff curve is
-  // evaluated, so the interaction band (especially the subduction mountain
-  // belt) does not trace the Voronoi polygon 1:1 — without this, coastlines
-  // on active margins are forced straight by the uniform +0.35 elevation
-  // band stamped along the plate edge.
-  const pass5Warp = new OpenSimplexNoise((nv.seed ^ 0xfa5170c5) | 0);
-  const warpFreqPass5 = nv.frequency * tv.coastlineWarpFrequency;
-  const npWarp = new Float32Array(3);
-
-  for (let i = 0; i < size; i++) {
-    const bi = nearestBoundary[i];
-    if (bi < 0) continue;
-
-    const b = boundaries[bi];
-    const pixelPlate = plateMap[i];
-    const isSubducting = b.subductingPlate === pixelPlate;
-
-    // Compute per-pixel falloff width based on interaction type and side
-    let multiplier = FALLOFF_MULTIPLIER[b.interactionType];
-    if (b.interactionType === InteractionType.Subduction && isSubducting) {
-      multiplier = SUBDUCTION_TRENCH_MULTIPLIER;
-    } else if (b.interactionType === InteractionType.OceanicConvergence && isSubducting) {
-      multiplier = SUBDUCTION_TRENCH_MULTIPLIER;
-    }
-    const effectiveFalloff = baseFalloff * multiplier * falloffScale;
-
-    // Noise-warp the distance so the interaction band zigzags instead of
-    // tracing the plate polygon. Amplitude is capped at 0.8·effectiveFalloff
-    // so the band can jitter but cannot stamp phantom mountains far inland.
-    const px = i % width;
-    const py = (i - px) / width;
-    sphericalEmbed3D(px, py, width, height, npWarp);
-    const warpSample = pass5Warp.eval3D(
-      npWarp[0] * warpFreqPass5,
-      npWarp[1] * warpFreqPass5,
-      npWarp[2] * warpFreqPass5
-    );
-    const warpLim = 0.8 * effectiveFalloff;
-    const rawOffset = warpSample * effectiveFalloff * tv.coastlineWarpAmplitude * 3;
-    const distOffset = rawOffset > warpLim ? warpLim : rawOffset < -warpLim ? -warpLim : rawOffset;
-    const warpedDist = Math.max(0, dist[i] + distOffset);
-
-    if (warpedDist >= effectiveFalloff) continue;
-
-    const t = 1 - warpedDist / effectiveFalloff;
-    const falloff = t * t; // quadratic for smoother edges
-
-    const plateType = plates[pixelPlate].type;
-    const { elevDelta, mountainRange } = interactionElevation(
-      b.interactionType,
-      plateType,
-      b.intensity,
-      falloff,
-      isSubducting
-    );
-
-    // Apply interaction delta to every pixel in the falloff band, including
-    // boundaries (t=1 there, matching the previously pinned crest magnitude).
-    baseElevation[i] += elevDelta;
-    mountainRanges[i] = mountainRange;
-    faults[i] = b.intensity * falloff;
-  }
-
-  return { baseElevation, faults, mountainRanges, continentalSubRelief, distToRidge, oceanAge };
-}
-
-/**
- * BFS from boundary pixels to propagate the nearest boundary index
- * outward to all pixels within falloff range.
- */
-function propagateBoundaryIndex(
-  nearestBoundary: Int16Array,
-  isBoundary: Uint8Array,
-  dist: Float32Array,
-  falloffWidth: number,
-  width: number,
-  height: number
-): void {
-  const queue: number[] = [];
-  const visited = new Uint8Array(width * height);
-
-  // Seed BFS with boundary pixels
-  for (let i = 0; i < isBoundary.length; i++) {
-    if (isBoundary[i]) {
-      visited[i] = 1;
-      queue.push(i);
-    }
-  }
-
-  let head = 0;
-  while (head < queue.length) {
-    const idx = queue[head++];
-    const x = idx % width;
-    const y = (idx - x) / width;
-
-    const neighborOffsets = [
-      y > 0 ? idx - width : -1,
-      y < height - 1 ? idx + width : -1,
-      x > 0 ? idx - 1 : y * width + width - 1,
-      x < width - 1 ? idx + 1 : y * width,
-    ];
-
-    for (const nIdx of neighborOffsets) {
-      if (nIdx < 0 || visited[nIdx]) continue;
-      if (dist[nIdx] >= falloffWidth) continue;
-
-      visited[nIdx] = 1;
-      nearestBoundary[nIdx] = nearestBoundary[idx];
-      queue.push(nIdx);
-    }
-  }
-}
-
-/**
- * Two-pass Chamfer distance transform.
- * Returns approximate Euclidean distance to nearest boundary pixel.
- */
-function chamferDistance(isBoundary: Uint8Array, width: number, height: number): Float32Array {
-  const size = width * height;
-  const dist = new Float32Array(size);
-  const INF = width + height;
-
-  for (let i = 0; i < size; i++) {
-    dist[i] = isBoundary[i] ? 0 : INF;
-  }
-
-  // Forward pass (top-left to bottom-right)
-  for (let y = 0; y < height; y++) {
-    for (let x = 0; x < width; x++) {
-      const idx = y * width + x;
-      const left = y * width + mod(x - 1, width);
-      if (dist[left] + 1 < dist[idx]) dist[idx] = dist[left] + 1;
-      if (y > 0) {
-        const top = (y - 1) * width + x;
-        if (dist[top] + 1 < dist[idx]) dist[idx] = dist[top] + 1;
-        const topLeft = (y - 1) * width + mod(x - 1, width);
-        if (dist[topLeft] + 1.414 < dist[idx]) dist[idx] = dist[topLeft] + 1.414;
-        const topRight = (y - 1) * width + mod(x + 1, width);
-        if (dist[topRight] + 1.414 < dist[idx]) dist[idx] = dist[topRight] + 1.414;
-      }
-    }
-  }
-
-  // Backward pass (bottom-right to top-left)
-  for (let y = height - 1; y >= 0; y--) {
-    for (let x = width - 1; x >= 0; x--) {
-      const idx = y * width + x;
-      const right = y * width + mod(x + 1, width);
-      if (dist[right] + 1 < dist[idx]) dist[idx] = dist[right] + 1;
-      if (y < height - 1) {
-        const bottom = (y + 1) * width + x;
-        if (dist[bottom] + 1 < dist[idx]) dist[idx] = dist[bottom] + 1;
-        const bottomLeft = (y + 1) * width + mod(x - 1, width);
-        if (dist[bottomLeft] + 1.414 < dist[idx]) dist[idx] = dist[bottomLeft] + 1.414;
-        const bottomRight = (y + 1) * width + mod(x + 1, width);
-        if (dist[bottomRight] + 1.414 < dist[idx]) dist[idx] = dist[bottomRight] + 1.414;
-      }
-    }
-  }
-
-  return dist;
-}
-
-// ── Main entry point ─────────────────────────────────────────────────────
-
-export function generateTectonicPlates(
-  width: number,
-  height: number,
-  nv: NoiseVariables,
-  tv: TectonicVariables
-): TectonicResult {
-  const { plateCount, cellCount, relaxationIterations } = tv;
-  const rng = { s: (nv.seed ^ 0xabcdef01) | 1 };
-
-  // Step 1: Spherical Voronoi tessellation. The cell grid is equirectangular
-  // (row 0 = north pole, lat = π/2 − (row+0.5)/H · π) and wraps in longitude
-  // exactly like the cylindrical case, so downstream raster consumers are
-  // unchanged. Seed positions are unit vectors on the sphere.
-  const sphereVoronoi = generateSphereVoronoi({
-    width,
-    height,
-    seedCount: cellCount,
-    seed: nv.seed,
-    relaxationIterations,
-  });
-  const { cells: cellMap, seeds: sphereSeeds } = sphereVoronoi;
-
-  // Project sphere seeds → pixel coords for the cylindrical seed-placement and
-  // boundary-vector helpers below. These helpers compare adjacent or near-by
-  // cells where tangent-plane (pixel) distance is a fine approximation; the
-  // sphere-correct work happens in centroid math (Step 5) and the cos(lat)
-  // scaling in classifyBoundaries (Step 6).
-  const seeds = sphereSeedsToPixels(sphereSeeds, width, height);
-
-  // Step 2: Build cell adjacency graph
-  const graph = buildVoronoiGraph(cellMap, width, height, cellCount, true);
-
-  // Step 3: Group cells into plates
-  const targetFractions = generateSizeDistribution(plateCount, rng);
-  const plateSeeds = selectPlateSeeds(seeds, cellCount, plateCount, width, targetFractions, graph.neighbors, rng);
-  const cellToPlate = groupCellsIntoPlates(graph.neighbors, cellCount, plateCount, plateSeeds, targetFractions);
-
-  // Step 4: Build pixel-level plate map
-  const plateMap = new Int32Array(width * height);
-  for (let i = 0; i < plateMap.length; i++) {
-    plateMap[i] = cellToPlate[cellMap[i]];
-  }
-
-  // Step 5: Compute plate centroids (averaging Vec3 unit vectors → renormalize
-  // → project to pixel) and assign properties.
-  const centroids = computePlateCentroidsSphere(sphereSeeds, cellToPlate, cellCount, plateCount, width, height);
-  const plates = assignPlateProperties(centroids, plateCount, width, height, nv.seed);
-
-  // Step 6: Aggregate and classify plate-level boundaries
-  const rawBoundaries = aggregatePlateBoundaries(graph.edges, cellToPlate, plateCount);
-  const boundaries = classifyBoundaries(rawBoundaries, plates, centroids, width, height);
-
-  // Step 7: Rasterize interactions into baseElevation, faults, mountainRanges,
-  // continentalSubRelief, distToRidge, and oceanAge.
-  const { baseElevation, faults, mountainRanges, continentalSubRelief, distToRidge, oceanAge } =
-    rasterizePlateInteractions(plateMap, boundaries, plates, width, height, nv, tv);
-
-  return {
-    baseElevation,
-    faults,
-    mountainRanges,
-    plateMap,
-    plates,
-    boundaries,
-    continentalSubRelief,
-    distToRidge,
-    oceanAge,
-  };
-}
